# gmmsales-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GMMSales API Documentation",
        "description": "Dokumentasi API untuk aplikasi GMMSales - sistem pencatatan kunjungan customer oleh sales lapangan dengan dokumentasi GPS dan foto.",
        "contact": {
            "name": "Rakha Hafizh",
            "email": "rakhahafizh@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "GMMSales Local Development Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Daftar semua produk",
                "description": "Mengambil seluruh daftar produk yang tersedia. Dapat diakses oleh semua user yang sudah login untuk keperluan dropdown saat mendaftarkan customer.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari produk berdasarkan nama",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products": {
            "post": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Tambah produk baru (Admin)",
                "operationId": "a19f84741edc4e6e9599d1d0ed8b3360",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_produk",
                                    "harga"
                                ],
                                "properties": {
                                    "nama_produk": {
                                        "type": "string",
                                        "example": "50 Mbps Internet"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 270000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil ditambahkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}": {
            "put": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Edit produk (Admin)",
                "operationId": "93fb5210f414ef18ce2972f108e980f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diupdate"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Hapus produk (Admin)",
                "operationId": "9167119fd9d754c7b03fb0afd13400f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum tanpa prefix 'Bearer ' (cukup tokennya saja)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}